Rename The Dev Grid to ASCII Table Pro
The Dev Grid is a project that Galin Iliev and I started some time ago.
Our idea was to have a developer portal with resources for professional
programmers, hobbiests and students. We had planned to start with the
ASCII Table and add more resources with time. We then abandoned the
project. It makes no sense not use what we've developed so far. I do
plan to add anymore funcionality for the moment. That's why I am
renaming it to ASCII Table Pro